Fix: setTextDatum引数をuint8_tに変更し、M5GFXの型安全性を確保
- IDisplay.hのsetTextDatum引数をintからuint8_tに変更
- DisplayAdapter.hでsetTextFont後にsetTextDatumを復元する修正
- 全てのテストファイルのモッククラスをuint8_tに対応
- メイン画面の現在時刻表示が中央寄せに修正

+++ lib/libaimatix/src/IDisplay.h
@@ -8,7 +8,7 @@ public:
     virtual void setTextColor(uint16_t color, uint16_t bgColor) = 0;
     virtual void fillRect(int x, int y, int w, int h, uint16_t color) = 0;
     virtual void drawRect(int x, int y, int w, int h, uint16_t color) = 0;
-    virtual void setTextDatum(int datum) = 0;
+    virtual void setTextDatum(uint8_t datum) = 0;
     virtual void setTextFont(int font) = 0;
     virtual void fillProgressBarSprite(int x, int y, int w, int h, int percent) = 0;
     virtual void drawLine(int x0, int y0, int x1, int y1, uint16_t color) = 0;
