+++ test/pure/test_input_display_view_pure/test_main.cpp
@@ -20,7 +20,7 @@ public:
     void setTextColor(uint16_t color, uint16_t bgColor) override { setTextColorCalled = true; }
     void fillRect(int x, int y, int w, int h, uint16_t color) override { fillRectCalled = true; }
     void drawRect(int x, int y, int w, int h, uint16_t color) override { drawRectCalled = true; }
-    void setTextDatum(int datum) override { setTextDatumCalled = true; }
+    void setTextDatum(uint8_t datum) override { setTextDatumCalled = true; }
     void setTextFont(int font) override { setTextFontCalled = true; }
     void fillProgressBarSprite(int x, int y, int w, int h, int percent) override { fillProgressBarSpriteCalled = true; }
     void drawLine(int x0, int y0, int x1, int y1, uint16_t color) override { drawLineCalled = true; }
